Adding module name to the end of generated symbols.
Currently,in the case when two symbols have the same name, they are
distinguished by appending an integer to the end of the symbol name. However
that does make it confusing while reading the code. The changes made here also
append the module name to the end of the generated symbols which makes the
code more readable.

Some points worth making a note:
* In case the symbols are still not unique (even after appending module name)
  then we append a unique integer to the end of the symbol for disambiguation.
* The moduleName has been appended only in the cases when actual
  collision occurs between the names.
* Currently, we consider renaming only functions and globals under our
  current scheme.